place applyFilters.js method in functions folder
