+++ frontend/src/pages/home/Home.js
@@ -2,11 +2,11 @@ import React, { useEffect, useState, useRef } from 'react'
 import JobCard from '../../components/jobCard/JobCard';
 import "./home.css"
 import { fetchData } from "../../utils/functions/fetchData"
+import { applyFilters } from "../../utils/functions/applyFilters"
+import FilterJob from '../../components/filterJob/FilterJob';
 
 //All the filter data options
 import { roles_Options, experience_Options, work_Options, salary_Options, location_Options } from '../../utils/data/FilterData';
-
-import FilterJob from '../../components/filterJob/FilterJob';
 
 const Home = () => {
 
@@ -57,56 +57,11 @@ const Home = () => {
 
     // Apply filters when filters or job data changes
     useEffect(() => {
-        applyFilters(jobData);
+        const filteredData = applyFilters(jobData, filters, searchTerm);
+        setFilteredJobData(filteredData);
+
     }, [filters, searchTerm, jobData]);
 
-
-    // Function to apply filters on job data
-    const applyFilters = (data) => {
-
-        let filteredData = data;
-
-        //Search (startsWith() used to check whether a string starts with a specific substring)
-        if (searchTerm) {
-            filteredData = filteredData.filter(job =>
-                job.companyName.toLowerCase().startsWith(searchTerm.toLowerCase())
-            );
-        }
-
-        // Filter by roles
-        if (filters.roles.length > 0) {
-            filteredData = filteredData.filter(job => filters.roles.includes(job.jobRole));
-        }
-
-        // Filter by experience
-        if (filters.experience.length > 0) {
-
-            filteredData = filteredData.filter(job => {
-                const selectedExperience = Math.max(...filters.experience);
-                return job.minExp >= selectedExperience
-            });
-        }
-
-        // Filter by work arrangement
-        if (filters.work.length > 0) {
-            filteredData = filteredData.filter(job => filters.work.includes(job.location));
-        }
-
-        // Filter by salary
-        if (filters.salary.length > 0) {
-            filteredData = filteredData.filter(job => {
-                const selectedSalary = Math.max(...filters.salary);
-                return job.minJdSalary >= selectedSalary;
-            });
-        }
-
-        // Filter by location
-        if (filters.location.length > 0) {
-            filteredData = filteredData.filter(job => filters.location.includes(job.location));
-        }
-
-        setFilteredJobData(filteredData);
-    };
 
     // Handle filter changes
     const handleFilterChange = (filterType, selectedValues) => {
